# apps/portfolio-service/openapi.yaml
openapi: "3.0.3"
info:
  title: Portfolio Service API
  version: 1.0.0
  description: Manages portfolios, positions, and risk results for RiskOps.

servers:
  - url: http://localhost:8081
    description: Via API Gateway
  - url: http://localhost:8082
    description: Direct to portfolio-service

paths:
  /api/portfolios:
    get:
      operationId: listPortfolios
      summary: List all portfolios
      tags: [portfolios]
      responses:
        "200":
          description: Array of portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createPortfolio
      summary: Create a new portfolio
      tags: [portfolios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioRequest"
      responses:
        "201":
          description: Created portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/portfolios/{id}:
    parameters:
      - $ref: "#/components/parameters/PortfolioID"

    get:
      operationId: getPortfolio
      summary: Get portfolio by ID
      tags: [portfolios]
      responses:
        "200":
          description: Portfolio details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deletePortfolio
      summary: Delete a portfolio
      tags: [portfolios]
      responses:
        "204":
          description: Portfolio deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/portfolios/{id}/positions:
    parameters:
      - $ref: "#/components/parameters/PortfolioID"

    get:
      operationId: listPositions
      summary: List positions for a portfolio
      tags: [positions]
      responses:
        "200":
          description: Array of positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: upsertPosition
      summary: Create or update a position
      tags: [positions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPositionRequest"
      responses:
        "200":
          description: Upserted position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/portfolios/{id}/positions/{symbol}:
    parameters:
      - $ref: "#/components/parameters/PortfolioID"
      - name: symbol
        in: path
        required: true
        schema:
          type: string
        description: Ticker symbol

    delete:
      operationId: deletePosition
      summary: Delete a position from a portfolio
      tags: [positions]
      responses:
        "204":
          description: Position deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/portfolios/{id}/risk/latest:
    parameters:
      - $ref: "#/components/parameters/PortfolioID"

    get:
      operationId: getLatestRisk
      summary: Get latest risk results for a portfolio
      tags: [risk]
      responses:
        "200":
          description: Array of latest risk results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/portfolios/{id}/risk:
    parameters:
      - $ref: "#/components/parameters/PortfolioID"

    get:
      operationId: getRiskHistory
      summary: Get risk result history for a portfolio
      tags: [risk]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
          description: Maximum number of results to return
      responses:
        "200":
          description: Array of risk results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    PortfolioID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Portfolio ID

  schemas:
    Portfolio:
      type: object
      required: [id, name, description, currency, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        currency:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreatePortfolioRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          default: ""
        currency:
          type: string
          default: USD

    Position:
      type: object
      required: [portfolio_id, symbol, weight, updated_at]
      properties:
        portfolio_id:
          type: integer
          format: int64
        symbol:
          type: string
        weight:
          type: number
          format: double
        updated_at:
          type: string
          format: date-time

    UpsertPositionRequest:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
          minLength: 1
        weight:
          type: number
          format: double

    RiskResult:
      type: object
      required: [id, portfolio_id, asof_date, horizon_days, alpha, method, metric, value, model_version, created_at]
      properties:
        id:
          type: integer
          format: int64
        portfolio_id:
          type: integer
          format: int64
        asof_date:
          type: string
        horizon_days:
          type: integer
        alpha:
          type: number
          format: double
        method:
          type: string
        metric:
          type: string
        value:
          type: number
          format: double
        model_version:
          type: string
        created_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"